+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Changed
 
 - Reduce zoom distance by half.
+- Removed damping on asteroids.
 
 ## [0.0.7]